Mapper simuleringsdata over i generasjoner/graphql-api

+++ sykepenger-api/src/main/kotlin/no/nav/helse/spleis/graphql/dto/GraphQLSimulering.kt
@@ -16,7 +16,7 @@ data class GraphQLSimuleringsdetaljer(
     val konto: String,
     val belop: Int,
     val tilbakeforing: Boolean,
-    val sats: Int,
+    val sats: Double,
     val typeSats: String,
     val antallSats: Int,
     val uforegrad: Int,
